feat(person): update school staff display, position labels, birthdate picker, grid borders, and pending approval table

+++ scripts/sync-workgroups-chainat.ts
@@ -50,7 +50,7 @@ async function main() {
     .orderBy(workgroups.sortOrder);
 
   const [total] = await db
-    .select({ n: sql<number>`count(*)::int` })
+    .select({ n: sql<number>`count(*)` })
     .from(workgroups);
 
   console.log(`sync เสร็จ — workgroups ใน DB: ${total?.n ?? 0}`);
